// modules/shared/types/about.ts
export interface About {
    mision: string;
    vision: string;
    moral: string[];
    experience: string;
    id: string;
    name: string;
    description: string;
}

export interface Contact {
    id: string;
    name: string;
    city: string;
    localLine: string;
    chatLine: string;
    email: string;
    address: string;
}